Translation to Spanish
Hi!

Here's the translation for SPANISH. Although I speak Spanish (Spain), I've made it as neutral as possible for the Latin American folks, as it differs a little in some expressions.

Thanks for the mod, and efforts. Truly enjoyed it. Here's my little grain of sand.

Have a nice day!